Add setup.py
setup.py was created using the template from
https://github.com/pypa/sampleproject/blob/master/setup.py.
Contents that may not be relevant to the gbasis have been removed, and contents
that may be needed in the future (but not now) have been commented out.

We may need to revisit the setup.py at one point to check if any crucial
information has been removed.

What:
1. Add setup.py
2. Add requirements.txt