Add package directories
What:
1. Add gbasis directory in the package parent directory
2. Add tests directory in the package parent directory

Issues:
1. I think there are good reasons to store gbasis should be inside an src
directory (import parity, ensure installation when testing, prevent import of
package in setup.py), it is not because I am still not super comfortable with
adding an extra directory
2. The choice betwen adding tests directory to the parent directory or in the
directory of the code boils down to whether tests should be installed alongside
the rest of the code. Though the developers are unaffected, the users will not
be able to test the code. However, separating out the tests will create a
cleaner separated environment for testing the code and can have dependencies
independent of the main package. I want to try separating it to see if one is
better.