Add derivative of Cartesian primitive
What:
1. Add function for evaluating the derivative of the cartesian primitive.
Multiple ways of evaluation are proposed.
2. Add function for evaluating the cartesian primitive.
3. Add tests using finite difference

Why:
1. There are potentially multiple ways of computing this derivative, but, I am
at the moment, unsure which would be better. Simple test seems to show that
there's not much of a difference, but the test set may have been a little small
2. Since the cartesian primitive needs to be evaluated anyways in order to
evaluate the derivative, we get these for free. The added function is simply a
wrapper around the evaluation of the derivative.

Issues:
1. Answers for the tests for the derivative (finite difference of the cartesian
primitive) depends on the code to evaluate the cartesiana primitive. Maybe it
would be better to have test answers that are independent of the code being
tested.