Disable warnings on pylint
What:
1. Disable W0511 (FIXME and other warning notes)
2. Disable C0330 (hanging indentation before block)

Why:
1. Seeing a this code is pre-alpha stage, FIXME's and other warning notes will
be critical for quickly keeping note of what needs to be done and fixed without
having to go through some sort of workflow. It is perfectly acceptable to write
down things to do/fix on the code to do in the future because we have higher
priority things to do. Not writing them to satisfy pylint seems unnecessary.
Nonetheless, the appropriate issue/card should be made in either github or
trello.
2. Hanging indentation is quite common using black. For easy standardization of
formatting, we keep black's format and ignore pylint's.