Change API for function that evaluates derivative
What:
1. Change name of function from eval_deriv_contraction to
_eval_deriv_contractions
2. Change name of argument from `coord` to `coords`
3. Remove eval_contraction
4. Move eval_deriv_prim, and eval_prim to tests

Why:
1. For higher level functions, it would be nice to work directly with
ContractedCartesianGaussians istances rather than their attributes. The
evaluation (and its derivatives) of the ContractedCartesianGaussians will be
used by all of the higher level so it seems feasible to change some of the
inputs of the `eval_deriv_contraction` to ContractedCartesianGaussians instance.
However this will hard-code the structure of the data class into this function,
which may cause problems if we decide to expand upon the
ContractedCartesianGaussians data class, such as the generalized contractions.
In anticipation of this change, we make a low level function whose inputs are
explicit numpy arrays rather than a data class.
2. Technically more than one coordinate can be evaluated at the same time.
3. Function for evaluating the ContractedCartesianGaussians instance will be
added later
4. My guess was that the primitives (and its derivatives) do not need to be
evaluated except for the testing purposes.