Add FIXME to lincomb.lincomb_blocks_evals
One of the parameters to `lincomb.lincomb_blocks_evals` is the function that
returns a numpy array. Though `lincomb.lincomb_blocks_evals` was designed in
anticipation of different functions, it seems that the main (if not only) use
case for `lincomb.lincomb_blocks_evals` is the evaluation of a contracted
gaussian and its derivatives. Then, maybe we can simply hard code the
function (probably `deriv.eval_deriv_shell` in.