Normalize primitives inside _eval_deriv_contractions
What:
1. Add norm to parameters
2. Change coordinates shape from `(3, N)` to `(N, 3)`
3. Remove support for primitive evaluation. Only input that follows the docstring is
supported.
4. Add an axis to all multidimensional arrays for the coordinate
5. Change ordering in nested for loop, remove one loop
6. Normalize primitives before transforming primitives to contractions
7. Make necessary changes to the tests and other dependent functions

Why:
1. Though the parameters already contain all the information needed to compute
the norm, norm is passed as a parameter because we assume that it will be
created by the data class, `ContractedCartesiangaussians`
2. This shape makes it easier to write down the coordinates. All it costs is a
transpose.
3. I am not sure how to feel about adding input checks. Since this is a low
level code, I am hoping that the code on top of this function will check/process
the input instead. This way, it should be easier to optimize this code, though
I'm not sure how much of a headache input check will cause. I think the two
choices are explicit support for a single type of input (via docstring) or some
sort of input processing that raises an error if something goes wrong.
4. This will prevent a bug in the future, I think. Since numpy arrays are
broadcast from the right, implicit dimensions on the right side will likely
cause problems when multiple coordinates are given.
5. Loop is reordered such that subsequent np.prod is done over axes that are
closer together (2, 4) vs (0, 4). Also, a for loop for the angular momentum
vector is removed because it can be vectorized at the expense of np.prod over
axes that are closer together (3, 4). Following code was used to time the code

```python
import timeit
print(
    timeit.timeit(
        '''
for k, l in it.product(range(3), range(3)):
orders = np.zeros(3, dtype=int)
orders[k] += 1
orders[l] += 1
for i in range(6):
    test = ContractedCartesianGaussians(
        5, np.array([0.5, 1, 1.5]), 0, np.random.rand(20), np.random.rand(20)
    )
    eval_deriv_shell(coords=np.random.rand(100, 3), orders=orders, shell=test)
        ''',
        setup='''
import itertools as it
from gbasis.contractions import ContractedCartesianGaussians
from gbasis.deriv import eval_deriv_shell
import numpy as np
        ''',
        number=1,
    )
)
```