Support generalized contractions when evaluating contraction and its derivative
Turns out the existing code is already compatible with generalized contractions.

1. Docstring changed to specify the API
2. Add tests

Multiple shapes of `prim_coeffs` are supported at the expense of concise API.
For example, we assumed that the `prim_coeffs` is always 2-dimensional, such
that the coefficients for a segmented contraction will have the shape `(K, 1)`.
However, it takes a little more time to evaluate a two-dimensional array over a
one-dimensional array, even if their size is the same. We support different
shapes of `prim_coeffs` to keep this little bit of performance.