Fix bug assigning tril blocks from triu blocks in BaseTwoIndexSymmetric
Each block within an overlap matrix (for cartesian and spherical) corresponds to
different contractions on the left side (row) and the right side (column). Since
the overlap is symmetric, we only need to compute the upper triangle blocks to
be able to construct the whole overlap matrix. To do so, I simply assigned the
tril blocks with the transpose of the corresponding triu blocks. However, I
didn't account for the fact that the ordering for the tril blocks is different
from the ordering of the triu blocks and I was using the same ordering for both
triu and tril blocks. Then, incorrect blocks are assigned to incorrect
positions, resulting in bad stack or wrong results.

To fix, blocks were ordered correctly and then assigned.

Issues:
We were unable to detect this bug in the tests because the overlap consisted of
only 4 blocks (2x2) which does not have this problem (same ordering for triu and
tril indices). The correct procedure is to add a better test (at least 9 blocks)
to catch this bug. However, these tests were quite painful to write (and I don't
want to do them) so I wrote some TODO's instead.