Add norm of the contractions
The given contraction coefficients are not necessarily normalized such that the
contractions are normalized (after normalizing the primitives). In order to
normalize it, different contractions (for different angular momentum vectors)
need to be multiplied with different factors. Since these factors are different,
it will be better to store it as an attribute rather than computing it each time.

Issues:
Since `overlap` module imports the `contractions` module, there is a cyclic
import problem if the `overlap` is imported at the start of the module. So the
import statement was moved to the method that calls the Overlap method.

Pylint still complains about the circular import, though.