Change index convention in _one_elec_int.py
Add an index for the "m" and move the index for the component vector to the
position 1.

The main point of setting these conventions is to prevent reindexing later on.
It seems that there is already an index convention for the integrals, line 56,
which seems to contradict the convention for the coordinates and exponents. This
commit makes the conventions consistent