Vectorize over increments of angular moment components in _one_elec_int
I'm not entirely sure that vectorizing these components does anything (timing
shows that the difference may be negligible), but they are vectorized
nonetheless. It is possible to vectorize other components too, but the indexing
gets way too complicated for them to be feasible, especially since there isn't a
significant improvement in performance.