Change ordering of the indices in horizontal recursion
Since indices for "a" spans more indices than the indices for "b", i.e. more
slicing, it may be faster to move the indices for "b" closer to the left. I
think in terms of accessing elements, finding specific elements is faster as the
index is closer to the left.

No noticeable performance improvements, however. Maybe the angular momentum is
too low to see anything significant (was tested at angular momentum 5)