Normalize the contractions in _compute_one_elec_integrals
The normalization constants for the primtives are different depending on the
angularr momentum of the contractions. Before the primitives are contracted,
they must be normalized (otherwise, the numbers keep growing with each
recursion).

The normalization constants must be hardcoded (rather than using the property
norm_prim in ContractedCartesianGaussians) because there are many more angular
momentums in the intermediate integral quantity than provided by the norm_prim
property.