Turn boys_func into an abstractstaticmethod
What:
1. Turn PointChargeIntegral.boys_func into an abstract static method
2. Turn PointChargeIntegral.construct_array_contraction into a class method
3. Remove boys_func as a parameter

Why:
1. Since boys_func is simply located in this class for organizational purposes
with no use for the PointChargeIntegral instances, it is turned into a static
method. Since this method determines the behaviour of its children, it is
turned into an abstract method.
2. This one is not that necessary, but the only dependency on the class
PointChargeIntegral is the static method, boys_func, this method is turned into
classmethod more for cleanliness. It wouldn't make a difference, I think, if
this method was a normal method.
3. Since boys_func is retrieved from the class, it no longer needs to be a
parameter.