Fix normalization constants in one electron integral
Issues:
It's still not fixed. Unable to fully reproduce HORTON's nuclear electron
attraction. In fact, the integral can be reproduced for a primitive of the s
type if we disable normalization altogether (but not the others). We can
reproduce almost all types of primitives if we only disable the normalization
associated with the exponents (and include the one with the angular momentum)
and if the exponent is 1. However, the f orbital (a=(1,1,1)) case seems to
break (zero for everything) so i need to fix that first