Remove if statement in base index classes
Since each matrix associated with the contracted Cartesian Gaussians is a two
dimensional matrix, we do not need to treat matrices differently regardless of
the number of rows. We can remove the if statement for matrices with one row
vector.

Removing this if statement might make it easier for numba (or some other jit's)
to optimize the code