Add construct_array_mix to BaseOneIndex
Added function is a pretty blatant copy of the construct_array_spherical. We can
abstract out pieces from construct_array_cartesian and construct_array_spherical
and piece them together in construct_array_mix, but it doesn't seem that
feasible in other base classes.

Also, it may be (a little) more difficult to optimize `construct_array_miz` due
if statements so maybe it'll be better to keep the `construct_array_spherical`
and `construct_array_cartesian` separated so that it is easier to optimize these
(spherical and cartesian) because these will be the main use cases.