Add check for density matrix shape when making electrostatic potential
Using `num_cart`, `num_sph`, and `num_seg_cont`, we can figure out the shape of
the density matrix needed in order to be compatible with the generated hartree
potential. Since hartree potential is the bottleneck of the code, it would be
better to check this early on rather than after constructing the hartree
potential.

What:
1. Change API of _electrostatic_potential_base
  - Make coord_type into a argument (from keyword argument)
  - Remove base_func from
2. Incorporate the point charge function into the _electrostatic_potential_base
rather than passing them in as arguments
3. Check for the density shape from the computed number of contractions
4. Check type of coord_types
5. Update tests and docstrings

Why:
1. It seems more obvious and clear, especially since the cartesian and spherical
coordinate types need to be sepraated now