Add recursion for integral over differential operator
Using the overlap integral helper, we can skip a few steps and just recurse over
the orders of differntiation.

Issues:
I couldn't make this recursion work for zeroth order differentiation for s
orbitals (raises IndexError). This means that this function cannot be used to
compute overlaps of s orbitals. I can get the zeroth order differentiation
supported for contractions with angular momentum greater than zero, but to make
this behaviour more transparent, none of the zeroth order differentiation is
supported.