Change name of the ContractedCartesianGaussians to GeneralizedContractionShell
It's about time we changed this monstrosity of a class name into another
montrosity. Contracted Cartesian Gaussians isn't quite applicable now because
this class also contains information on the spherical contractions. Also, we now
support generalized contraction. These collections of contractions are often
called "shell", but that is confusing with the traditional usage of "shell" in
chemistry. To make these terms as clear as possible (with little contradiction
with the literature), we call the collection of contractions this class
describes as "generalized contraction shell".

What:
1. Change name of the class from ContractedCartesianGaussians to
GeneralizedContractionShell
2. Update changes everywhere
3. Add docstring clarifying  what "generalized contraction shell" means.