Remove charge attribute from GeneralizedContractionShell
Originally the charge attribute was thought to be needed in many calculations,
but it was only needed in electrostatic potential and the nuclear electron
attraction integral. In these cases, the nuclear charges (and coordinates) are
passed in explicitly and there is no need for the contraction object to contain
the nuclear charge information. In fact, it is never used.

1. Remove charge attribute in GeneralizedContractionShell
2. Remove setters and getters
3. Remove charges attribute in make_contraction
4. update Update everything (intializers and tests)