Select appropriate angmom components when compute two electron integrals
Before, there was a memory issue when computing the two electron integrals of
primtives at angular momentum 4 due to the construction of an array with at
least 4**12 elemens. To prevent the construction of this array, we select only
the angular momentum components that are needed after the horizontal recursion
over index d and b. Doing so, we have a much smaller array.

This lets us go up to angular momentum 6 with around 12 GB of memory. This is
still a problem.