Split horizontal recursion in two-electron integrals into two parts
Before the horizontal recursion, we create an array that can store all of the
results of the recurison. Since the horizontal recursion occurs over three
dimensions (x, y, z), the array includes the shape of max_angmom**3. For smaller
angular momentum, this is not a problem, but once we hit around angular momentum
of 6, this costs too much memory. Keep in mind that there are also
indices/dimensions for the angular momentum components of the other three
contractions as well as the segmented contractions of all four (generalized)
contractions.

To minimize memory usage, we split apart the horizontal recursion into two
parts: first part recurses over the first and second index (x and y) and the
second recurses over the third index (z). The trick is to select only the desired
angular momentum components in the x and y direction before constructing the
array for the second horizontal recursion.

The horizontal recursion was split off into two components rather than
three (one for each x, y, z) because the number of different angular momentum
components scales faster than the angular momentum. For example, if we have
angular momentum of 6, then we have 28 different angular momentum components.
Since we need to allocate at least one dimension for the angular momentum
components. Once we can replace two dimensions, i.e. 36 elements, we see
benefits of replacing them with the angular momentum components.

Note that this implementation still does not scale. Since all results of the
recursions are explicitly stored, we have an explosion of the number of
intermediates that are computed. At the moment, we see a limit of around angular
momentum 7 for a single segmented contraction.