Split zero angular momentum case out of _compute_two_elec_integrals
The electron transfer recursion has a term that transfers the angular momentum
from contraction a to c. However, when both contractions have angular momentum
of zero, the corresponding term goes to zero. In order to avoid an if statement,
we use slicing to utilize arrays with zero elements. For example the slice 1:2
would have no elements if there is only one element in this dimension (i.e.
angular momentum zero).

At the moment, it doesn't seem possible to assign array with zero elements to
another array with zero elements if the shapes do not match up. Or if there are
more than one 1-0 pairing.

Instead of creating an if statement, we create a special function that handles
specifically the case where all of the angular momentum of the contractions are
zero. There is a bit of duplicated code, but separating out this code should (I
think) make it esier for w/e optimizer to optimize the code. Treating the
special case separately will be trivial at the higher level.

1. Add function _compute_two_elec_integrals_angmom_zero
2. Remove support for zero angular momentum in _compute_two_elec_integrals
3. Add tests
4. Update docstring