Add support for spherical s and p orbitals in iodata
Iodata does not explicitly write out the spherical conventions for the s and p
orbitals, probably because they are easily converted into cartesian forms. I
understand that the s orbital is trivial to convert, but the the p orbitals
change ordering depending on the format. For example, if the ordering is by the
magnetic quantum number, then it would be x, z, y (or the reverse ordering).
However alphabetical ordering would result in x, y, z. I guessed that unless the
spherical ordering is given for the p orbitals, the cartesian ordering is used.

1. Add s orbital spherical orbital ordering convention in iodata wrapper
2. Add p orbital spherical orbital ordering convention in iodata wrapper
3. Add test