Ensure proper cancellation in real solid harmonic calculations
The dictionary for real_solid_harmonic was built assuming that every
element would only show up once, and thus coefficients were assigned to
the dict. This has been fixed by using a defaultdict instead, which
defaults to 0.0, allowing for terms to be cancelled out without adding
too many additional terms later on when the transformation matrix is
constructed.