Do not compute transform when it is not used.
This fixes a part of #78. In BaseOneIndex.construct_array_mix, the
transformation matrix from Cartesian to Spherical functions was
also computed for Cartesian shells. This caused some unexpected
issues when computing densities for wavefunction in Cartesian
basis sets, for which no ordering of the spherical functions
was defined. With this fix, also some unused calculations are
avoided.