Fix bug for assembling generalized contractions from gbs files
Fix bug in parsers.py which caused it to fail to load Pople 6-31's family of basis
sets as well as SVP Dunning-Hay and possibly more. By possible bug in np.allclose()
it worked on Pople family of STO-nG as well as Dunning basis sets such as aug-cc-pVTZ.

issue: np.allclose() compares two arrays element by element.
The purpose was to test if previous shells exponents were the same as current shells,
if True, hstack coefficients, else, append. np.allclose requires arrays to be same length,
however, it failed to raise error if one array was only length 1 (this was the fluke?).
In Pople STO-3G and Dunning, shells of same angular momentum were either same length, or
the following shell was length one i.e., single primitive. This is not the case in
Dunning-Hay, 6-31G, ANO and others. In these cases np.allclose returned error of mismatch
in array lengths.

I have added a patch to make sure arrays are same length, if not, don't stack coefficients.
If yes, test with np.allclose() to see if we hstack coefficients, or append.