Fix bug due to commit  908010e
In commit 908010e code in evaluate_deriv_basis() was changed.
This was due because in theory there was a bug with how the
function decides to choose between construct_array_cartesian,
construct_array_spherical and  construct_array_mix (in base_one.py).
It evaluates coord_type variable to be equal to either “cartesian” or
“spherical” and if this is False goes to  construct_array_mix. Problem is that
coord_type is sometimes a list or a tuple that can have all the items
equal to cartesian or spherical and this won’t enter the
construct_array_cartesian function. Code in commit  908010e
was trying to fix this but then tests
test_evaluate_deriv_basis_cartesian()
and test_evaluate_deriv_basis_spherical() in test_eval_deriv()
fail because they are passing coord_type argument
as “cartesian”/”spherical” strings and with the new  code
it went directly to mix because it evaluated wrong the string
 and then inside construct_array_mix it will raise an error
because this is a string and not a list/tuple . This is a basic fix
but I am sure a more pythonic/concise if statement can be used here.