Link contributing guidelines to QC-Devs code of conduct
Link to the web site for the central code of conduct, which is more up to date. I retained the old code of conduct as a file, just to be clear that we *have* one, but the code of conduct in QC-Devs is the one that is maintained and (periodically) updated.