Fix incorrect condition for skipping Libcint tests
The current test won't work because it only checks if the build files exist:
https://github.com/theochem/gbasis/blob/3a88b8400d17070e2e41e43ab281dea5a109279a/tests/test_libcint.py#L62-L63

It should actually check if a file matching `gbasis/integrals/lib/libcint.so*` exists, as this is the actual file that will be loaded, and included with gbasis when it is packaged.

This commit fixes the check, so that the libcint tests will only be skipped if Libcint is missing from the actual package.