Change API for screening overlap integrals
The arguments for screening overlap integrals had to be provided when making a basis
(i.e., list of contractions). Then, when basis was passed to the `overlap_integrals`
function, these arguemnts were checked. This did not make sense, and meant when adding
screening feature for other integrals, additional arguments had to added to the
`make_contractions` function. This API is changed so that screening is requested when
calling the `overlap_integrals` function.