Fix compatibility with changes in IOData (prepare_segmented)
Fixes #195

The required function is imported locally from the iodata library
because the wrapper module also contains a wrapper for pyscf.
When the import is made at the top and IOData is not installed,
it would also become impossible to use the pyscf wrapper.
(You may want to split the wrapper module into one per package
to avoid local imports, but this is not critical.)

This commit includes a few side effects:

- Ruff fixes were required in wrappers.py to make the commit.
- .envrc is added to .gitignore. I used direnv for managing environments